Rename the generator and config file names
