--- lib/rails/generators/dynamo_db_session_handler/USAGE
@@ -1,12 +1,0 @@
-Description:
-    Generates a migration file for deleting and a creating a DynamoDB
-    sessions table.
-
-Example:
-    rails generate dynamo_db_session_migration <MigrationName>
-
-    This will create:
-        db/migrate/VERSION_migration_name.rb
-
-    This file will be migrated when the command rake db:migrate is run
-    in the command line.
