+++ lib/rails/generators/sessionstore/dynamodb/templates/sessionstore/USAGE
@@ -1,0 +1,13 @@
+Description:
+    Generates a migration file for deleting and a creating a DynamoDB
+    sessions table, and a configuration file for the session store.
+
+Example:
+    rails generate sessionstore:dynamodb <MigrationName>
+
+    This will create:
+        db/migrate/VERSION_migration_name.rb
+        config/sessionstore/dynamodb.yml
+
+    The migration will be run when the command rake db:migrate is run
+    in the command line.
