+++ lib/rails/generators/sessionstore/dynamodb/templates/sessionstore/dynamodb.yml
