+++ lib/rails/generators/sessionstore/dynamodb/templates/sessionstore_migration.rb
